[#103] Cleanup messaging pattern and move it into static config
