+++ iceoryx2/src/service/builder/publish_subscribe.rs
@@ -19,8 +19,8 @@ use std::marker::PhantomData;
 use crate::service;
 use crate::service::dynamic_config::publish_subscribe::DynamicConfigSettings;
 use crate::service::header::publish_subscribe::Header;
-use crate::service::messaging_pattern::MessagingPattern;
 use crate::service::port_factory::publish_subscribe;
+use crate::service::static_config::messaging_pattern::MessagingPattern;
 use crate::service::*;
 use iceoryx2_bb_elementary::alignment::Alignment;
 use iceoryx2_bb_log::{fail, fatal_panic, warn};
